Review and refactor C bindings
